+++ src/module/applications/sidebar/tabs/combat-tracker.mjs
@@ -1,5 +1,5 @@
-import { systemID, systemPath } from "../../../constants.mjs";
-import { DrawSteelCombatant, DrawSteelCombatantGroup } from "../../../documents/_module.mjs";
+import { systemPath } from "../../../constants.mjs";
+import { DrawSteelCombatant, DrawSteelCombatantGroup, DrawSteelTokenDocument } from "../../../documents/_module.mjs";
 
 /**
  * @import { ContextMenuEntry } from "@client/applications/ux/context-menu.mjs";
@@ -360,25 +360,40 @@ export default class DrawSteelCombatTracker extends sidebar.tabs.CombatTracker {
 
   /** @inheritdoc */
   _getCombatContextOptions() {
-    const entryOptions = super._getCombatContextOptions();
-
-    if (game.combats.isDefaultInitiativeMode) {
-      entryOptions.findSplice(o => o.name === "COMBAT.RollAll");
-      entryOptions.findSplice(o => o.name === "COMBAT.RollNPC");
-
-      entryOptions.unshift({
+    const entryOptions = [
+      {
         name: game.i18n.format("DOCUMENT.Create", { type: game.i18n.localize("DOCUMENT.CombatantGroup") }),
         icon: "<i class=\"fa-solid fa-users-rectangle\"></i>",
         callback: () => DrawSteelCombatantGroup.createDialog({}, { parent: this.viewed }),
       }, {
+        name: "DRAW_STEEL.CombatantGroup.GroupSelected",
+        icon: "<i class=\"fa-solid fa-users-viewfinder\"></i>",
+        callback: async () => {
+          /** @type {DrawSteelTokenDocument[]} */
+          const tokens = canvas.tokens.controlled.map(t => t.document);
+          await DrawSteelTokenDocument.createCombatants(tokens);
+          const combatants = tokens.map(t => t.combatant);
+          const actorName = tokens[0]?.actor.name;
+          const tokenImage = tokens[0].texture.src;
+          const type = tokens.some(t => t.actor?.system.isMinion) ? "squad" : "base";
+          const group = await DrawSteelCombatantGroup.create({
+            type,
+            name: tokens.every(t => t.actor?.name === actorName) ? actorName : DrawSteelCombatantGroup.defaultName({ type, parent: this.viewed }),
+            img: tokens.every(t => t.texture.src === tokenImage) ? tokenImage : null,
+          }, { parent: this.viewed });
+          const updateData = combatants.map(c => ({ _id: c.id, group: group.id }));
+          await this.viewed.updateEmbeddedDocuments("Combatant", updateData);
+          if (group.type === "squad") await group.update({ "system.staminaValue": group.system.staminaMax });
+        },
+      }, {
         name: "COMBAT.InitiativeRoll",
         icon: "<i class=\"fa-solid fa-dice-d10\"></i>",
+        condition: () => game.combats.isDefaultInitiativeMode,
         callback: () => this.viewed.rollFirst(),
-      });
+      },
+    ];
 
-    }
-
-    return entryOptions;
+    return entryOptions.concat(super._getCombatContextOptions());
   }
 
   /* -------------------------------------------------- */